Use mhtml class.
The idea is that we will include a standalone representation (wrapped
in a DIV) of a mizar article as part of a web page.  I want to use
Josef's hard-earned awesome HTMLization, but I don't want the CSS for
laying out the semantic representation of an article to be the way
my *entire* page is laid out.  To achieve this separation, I will
sandbox the elements that are generated by this code by assigning them
to the new class "mhtml".