Try to bake in an absolute path.
It seems that when the open method, when called on objects of the
class XMLHttpRequest, interprts URIs relative to where the souce
document comes from.  Thus, when visiting the page

  http://mizar.cs.ualberta.ca:4242/article/toler_1

and following an ajax'y link there to expand a proof results in the request

  GET /article/proofs/toler_1/ckb32/1

which is bad.  I want the request URI to be

   /proofs/toler_1/ckb32/1

not

  /article/proofs/toler_1/ckb32/1

I'm hoping that adjusting the URI in the onclick attribute of the
anchor elements will make a difference.