Go hunting for a term appearing in a dependent type that is not visible.
E.g., from RELSET_1:

definition
  let X,Y,R,A;
  redefine func R.:A -> Subset of Y;
  ...
end;

where

  A : set
  X : set
  Y : set
  R : Relation of X,Y

we need to determine that X is needed, even though it doesn't appear
in the context or explicitly in the other loci varibles.  It occurs as
an argument of a dependent type of something that is easily seen to be
needed (namely, R).