Some work on issue #5.
The article XREAL_0 illustrates the toplevel consider statement problem.

======================================================================
consider x1,x2,y1,y2 being Element of REAL such that
Lm9: 1 = [*x1,x2*] and
Lm10: -1 = [*y1,y2*] & 0 = [*+(x1,y1),+(x2,y2)*] by XCMPLX_0:def 4;
======================================================================

Here are two sketches of a solution, by hand (i.e., not using XSLT):

* xreal_0
* xreal_0-skolem

The latter one doesn't work.  It is based on a careful though
overengineered solution that introduces a new functor x2 depending on
x1, a a new functor y1 depending on x1 and y1, etc.  This turns out to
be unacceptable because we want to work with these variables
independent of their context.  y1 doesn't really depend on x1, and so
it goes for the other variables.

The first sketch does seem to work.

There are various stages of the process:

* For each variable v introduced by a toplevel choice statement, define

  - a new mode M(v), defined as a subtype of whatever type v has;

  - a zero-ary functor F(v) which is simply defined as equal to "the
    M(v)".

  Both of these steps involve making new vocabulary files.

* If there are any conditions that the new choice variables are
  assumed to have, make a new article that contains the conjunction of
  all these conditions; the justification is the final definition of
  the last step above;

  - For each condition (conjunct of the conjunction of conditions),
    make a new article containing only this condition; the
    justification is the conjunction.

* Toplevel items occurring later in the source article are then
  rewritten to appeal to these new conditions.